requirements.txt
Pillow==7.2.0
webbrowser==0.10.3
os==0.1